struct uuid_time: Use integer types guaranteed by the C99 Standard
Change the types in `struct uuid_time` to sizes guaranteed by C99:

  - hi, mid: unsigned int → unsigned short
  - low: unsigned int → unsigned long

Hasn't been a problem on 64-bit architecture, but there's some
complaining from compilers on 32-bit archs. All tests did succeed,
though.

6ac8e10e-d4f7-11eb-8e0f-4f45262dc9b5

+++ src/uuid.h
@@ -29,9 +29,9 @@
 typedef unsigned long long utime_t;
 
 struct uuid_time {
-	unsigned int hi;
-	unsigned int mid;
-	unsigned int low;
+	unsigned short hi;
+	unsigned short mid;
+	unsigned long low;
 };
 struct uuid {
 	struct uuid_time time;
